planner: move physical PhysicalLock related logic into physical op dir.
