+++ pkg/planner/core/generator/plan_cache/BUILD.bazel
@@ -5,7 +5,10 @@ go_library(
     srcs = ["plan_clone_generator.go"],
     importpath = "github.com/pingcap/tidb/pkg/planner/core/generator/plan_cache",
     visibility = ["//visibility:private"],
-    deps = ["//pkg/planner/core"],
+    deps = [
+        "//pkg/planner/core",
+        "//pkg/planner/core/operator/physicalop",
+    ],
 )
 
 go_binary(
