+++ pkg/planner/core/operator/physicalop/BUILD.bazel
@@ -10,6 +10,7 @@ go_library(
         "physical_hash_agg.go",
         "physical_index_join.go",
         "physical_limit.go",
+        "physical_lock.go",
         "physical_max_one_row.go",
         "physical_mem_table.go",
         "physical_merge_join.go",
